Revert "Fix dnlib API"
This reverts commit 940bb41b89c750e2469ac6d1ec72f50b48a19042.

+++ de4dot.code/deobfuscators/MyPEImage.cs
@@ -53,7 +53,7 @@ namespace de4dot.code.deobfuscators {
 
 		public ImageSectionHeader FindSection(RVA rva) {
 			foreach (var section in peImage.ImageSectionHeaders) {
-				if (section.VirtualAddress <= rva && (uint)rva < (uint)section.VirtualAddress + Math.Max(section.VirtualSize, section.SizeOfRawData))
+				if (section.VirtualAddress <= rva && rva < section.VirtualAddress + Math.Max(section.VirtualSize, section.SizeOfRawData))
 					return section;
 			}
 			return null;
